+++ changelog.md
@@ -14,7 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Fixed
 
-- Doc only: Correct description of rendering difference between `Some(EdgeAnchors(1))` and `None` (``)
+- Doc only: Correct description of rendering difference between `Some(EdgeAnchors(1))` and `None` (`caaabc4`)
+- Add automated straight-connector fixing-point test (``)
 
 # [Released]